Stop recording secrets on service account
A token for a service account can be generated by creating a secret for
that service account with the annotation
`kubernetes.io/service-account.name` pointing to that service account.

In Kubernetes 1.23 and below, a reference to the secret was
automatically stored directly on the service account by Kubernetes,
making it easy to look up. For Kubernetes 1.24, Rancher started
explicitly adding the secret to the service account so that it could
refer back to it.

In k8s 1.27, storing references to the secret is discouraged[1] in favor
of explicitly using the TokenReview API[2]. This does not serve
Rancher's purposes because the service account token needs to be
long-lived and is not intended to be mounted in a pod. To avoid the
warning, this change stops storing the secret reference on the service
account in favor of adding a label to the secret to make it easily
indexable.

This change affects these service accounts:

- cattle-token (the admin service account that Rancher uses to connect
  to the downstream)
- impersonation
- monitoring v1
- kontainer-engine
- CAPR machine bootstrap account
- v3.APIService service accounts

Some changes are made to the serviceaccounttoken package:

- EnsureSecretForServiceAccount is updated to use a wrangler cache
  instead of a lambda, so the norman secret lister can no longer be used
  with it. If the wrangler cache cannot be provided (as in the case of
  the cluster agent), a client is used instead.
- ServiceAccountSecretName is changed to ServiceAccountSecret and
  returns the entire secret instead of just the name.
  - Attempting to retrieve the whole secret no longer makes sense in a
    watch context, so the CAPR configserver package now just uses the
    new label to watch for the secret.
- If an invalid secret is found, it is deleted with a warning.

[1] https://github.com/kubernetes/kubernetes/blob/5ebb9f319576303cabc0216a8df536b0b4307e67/pkg/serviceaccount/legacy.go#L155
[2] https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/#manually-create-an-api-token-for-a-serviceaccount